move spacegraphcats.index functionality into spacegraphcats.cdbg
