+++ spacegraphcats/conf/Snakefile
@@ -204,7 +204,7 @@ rule make_contigs_kmer_index:
         join(catlas_dir, "contigs.fa.gz.mphf"),
         join(catlas_dir, "contigs.fa.gz.indices"),
      shell:
-        "{python} -m spacegraphcats.index.index_contigs_by_kmer -k {ksize} {catlas_dir}"
+        "{python} -m spacegraphcats.cdbg.index_contigs_by_kmer -k {ksize} {catlas_dir}"
 
 ### Search rules.
